+++ crates/circuits/src/bin/parse.rs
@@ -4,6 +4,8 @@ use std::fs::write;
 fn main() {
     build_aes();
     build_sha();
+    #[cfg(feature = "blake3")]
+    build_blake3();
 }
 
 fn build_aes() {
@@ -18,4 +20,12 @@ fn build_sha() {
 
     let bytes = bincode::serialize(&circ).unwrap();
     write("circuits/bin/sha256.bin", bytes).unwrap();
+}
+
+#[cfg(feature = "blake3")]
+fn build_blake3() {
+    let circ = mpz_circuits::circuits::blake3::compress();
+
+    let bytes = bincode::serialize(&circ).unwrap();
+    write("circuits/bin/blake3.bin", bytes).unwrap();
 }
